Implemented RJUMP instruction

+++ src/typedefs.h
@@ -9,8 +9,8 @@ typedef struct
   // Memory
   uint8_t PROGRAM_MEMORY[PROGRAM_MEMORY_SIZE];
   uint8_t SRAM[SRAM_SIZE];
-  uint8_t PC; // Program counter
-  uint8_t SP; // Stack pointer
+  uint32_t PC; // Program counter
+  uint32_t SP; // Stack pointer
 
   uint64_t tick_counter;
